(PC-18448)[PRO] refactor: link subrouters to route declarations
